+++ src/frontend/src/optimizer/plan_node/mod.rs
@@ -649,7 +649,7 @@ mod logical_join;
 mod logical_limit;
 mod logical_multi_join;
 mod logical_now;
-mod logical_over_agg;
+mod logical_over_window;
 mod logical_project;
 mod logical_project_set;
 mod logical_scan;
@@ -730,7 +730,7 @@ pub use logical_join::LogicalJoin;
 pub use logical_limit::LogicalLimit;
 pub use logical_multi_join::{LogicalMultiJoin, LogicalMultiJoinBuilder};
 pub use logical_now::LogicalNow;
-pub use logical_over_agg::LogicalOverWindow;
+pub use logical_over_window::LogicalOverWindow;
 pub use logical_project::LogicalProject;
 pub use logical_project_set::LogicalProjectSet;
 pub use logical_scan::LogicalScan;
